+++ HISTORY.md
@@ -21,6 +21,7 @@
         * new "expression" selector in block-attribute (metaprogramming) reporter
         * enables expressions (unringed blocks) to be CALLed and RUN
         * enabled unringed blocks to be shown in and dragged out of speech and result balloons
+    * drawing on sprites
     * files
         * new general "dta_export(data, name, type)" extension (for csv, json, etc.)
         * new general "dta_import(raw?)" extension (for text files, csv, json etc.)
@@ -42,6 +43,9 @@
     * fixed some special cases for using "combine" on an empty list, thanks, @rmunn!
 * **Translation Updates:**
     * German
+
+### 2025-12-05
+* objects: support drawing on sprites
 
 ### 2025-12-03
 * objects, gui: let tutorials manage data draggability through their scene setting